Connect to BLL and return Car List.

+++ SOAAssignment/Models/Car.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Web;
 
 namespace SOAAssignment.Models
@@ -14,11 +15,11 @@ namespace SOAAssignment.Models
         public string Model { get; set; }
         public int LicenceAge { get; set; }
         public int DriverAge { get; set; }
-        public int DailyMaxKm { get; set; }
-        public int CurrentKm { get; set; }
+        public double DailyMaxKm { get; set; }
+        public double CurrentKm { get; set; }
         public AirgBagEnum HasAirBag { get; set; }
         public int LuggageVolume { get; set; } // litre
         public int NumSeats { get; set; }
-        public int RentPrice { get; set; } // ₺
+        public decimal RentPrice { get; set; } // ₺
     }
 }
